vkbd works for tinyboot+guest CertiKOS, but not for guest grub

+++ sys/virt/svm/svm_handle.h
@@ -8,7 +8,7 @@
 #include <sys/virt/vmm.h>
 
 void svm_inject_event(struct vmcb *,
-		      uint8_t type, uint8_t vector, bool ev, uint32_t errcode);
+		      uint32_t type, uint8_t vector, bool ev, uint32_t errcode);
 
 void svm_guest_handle_gpf(struct vm *, tf_t *);
 int svm_guest_intr_handler(struct vm *, uint8_t irq);
